Update basicTextGame.js
Updates to the display for player health, and adding in a basic enemy generator.

Next steps:

CSS: Tidy up some of the boxes so they are a bit more static. The names of either the enemy or player profession stretch and contact their containers. Ideally want to give them a minimum width that is larger than the longest name, or at least add in a text wrap so it goes to another line.

JS: Use the existing random number generator to turn it into an attack, with the generator number being the damage dealt. Subtract that number from the enemy health, and have it so the enemy is "removed" when their health is 0 or less.