Tidying up part 1
Tidying up the bulk of the code; placing all starting variables at the top of the file, followed by DOM/HTML/CSS imports into JS, followed by functions. 

Also reworking the combat so it's a single parent function that gets invoke, that itself invokes multiple functions. RIght now it's not much of a difference, but eventually there might be numerous functions invoked every round of combat/combat end. At that point it would get unwieldly to rewrite the eventListener for every one of these. Instead, the goal here is to just add a new line to the "combat" function, type in the function name, and be done. Overall it will half or so the amount of typing needed, and make it cleaner to understand/look at.